backend/requirements.txt
flask==2.2.5
google-cloud-storage==2.14.0
openai==1.2.4
anthropic==0.49.0
cohere==4.25
vertexai>=1.43.0
google-auth==2.23.3
google-cloud-secret-manager==2.16.0
